backend/requirements.txt
uvicorn
fastapi
python-multipart
pydantic
ultralytics
torch
torchvision
opencv-python
opencv-python-headless
Pillow
pymongo
python-dotenv
numpy
python-magic
prometheus-client
slowapi
pyyaml
reportlab
colormath